Implement dithering and noise shaping
Dithering lowers digital-to-analog conversion ("requantization") error,
lowering distortion and replacing it with a constant, fixed noise
level, which is more pleasant to the ear than the distortion. Doing so
can with a noise-shaped dither can increase the dynamic range of 96 dB
CD-quality audio to a perceived 120 dB.

Guidance: experts can configure many different configurations of
ditherers and noise shapers. For the rest of us, these are sane
defaults depending on the output format:

| Format     | Ditherer         | Noise shaper             |
+------------+------------------+--------------------------+
| S16        | Triangular (tri) | Wannamaker (fw9 or fw24) |
| S24, S24_3 | High Pass (hp)   | None                     |
| S32        | None             | None                     |
| F32        | Not supported    | Not supported            |

Notes:

 * Try swapping ditherers with Gaussian if you prefer a more analog
   sound.

 * Try swapping noise shapers with Fraction Saving if you are running
   on power-constrained hardware. In this case, disable dithering.