Update change log
Update to document potentially breaking behaviour in player.rs, although currently none of the backend would be affected because they just panic! on error and don't actually return errors and/or the functions are noops in the case of start and stop. ALSA is the exception as it does actually implement all of them and returns errors on failure as it should as of this PR.