Add release-dist-optimized profile
release-dist-optimized inherits from `release`. Useful if you're distributing librespot at part of a project. The diffrences are:

panic = "abort", Makes librespot abort instead of unwind and hang on a panic.
Extremely useful when running librespot unattended as a system service for example to allow for auto-restarts.

codegen-units = 1 and lto = true, Take slighly longer to compile but produce more optimized binaries.