discovery::server: fix activeUser field of getInfo
The current active user was almost persisted and reported but
wasn't. The handler for the discovery server is under an Arc so I
introduced a Mutex to protect the username field and made setting the
username and sending the credentials atomic so that every valid
credential send corresponds to an active user update.