discovery: make features additive
i.e. add with-libmdns instead of using not(with-dns-sd). This is in
preparation for adding another backend, and in general recommended, cf.
https://doc.rust-lang.org/cargo/reference/features.html#mutually-exclusive-features

The logic is such that enabling with-dns-sd in addition to the default
with-libmdns will still end up using dns-sd, as before.
If only with-libmdns is enabled, that will be the default.
If none of the features is enabled, attempting to build a `Discovery`
will yield an error.