chore: Add ability to change player sample rate
Adds the ability for the player to change its sample rate, to be able to
accurately play back local files that aren't sampled at 44.1 kHz.

The functionality to change sample rate must be implemented individually
by each sink. In this commit, support for this has only been added to
the alsa sink.

Testing this by swapping between a 48.0 kHz local file and a 44.1 kHz
Spotify song appears to play back both without any distortion of sound.